+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Updated
 - Updated copy of "View this item elsewhere" buttons. (DR-3781)
 
+### Updated
+- updated middleware to strip /book from item urls
+
 ## [0.4.4] 2025-07-01
 ### Added
 - Added Playwright tests to confirm featured section on the homepage (DR-3658)